pyproject.toml
[tool.poetry]
name = "wyoming-whisper"
version = "0.1.0"
description = ""
authors = ["Simon Guigui <simon.guigui@parkoview.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai-whisper = "20231117"
wyoming = "^1.2.0"
tokenizers = "0.13.*"
black = "^23.11.0"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
